Opt: re-factor shop module; move buy select and amount operations to baseline classes, common currency variable as all shop variants are created independently
